deleted graph options, renamed chosen option
