+++ src/pages/Home.tsx
@@ -3,9 +3,7 @@ import Map from '../components/Map/Map';
 import {Text, Box, Flex, Spacer} from '@chakra-ui/react';
 import AqiDial from '../components/AqiGauge/AqiDial';
 import {useTranslation} from 'react-i18next';
-import AqiGraphScatter from '../components/AqiGraph/AqiGraphScatter';
-import AqiGraphLine from '../components/AqiGraph/AqiGraphLine';
-import AqiGraphScatterStack from '../components/AqiGraph/AqiGraphScatterStack';
+import AqiGraph from '../components/AqiGraph/AqiGraph';
 
 /**
  * Home screen component
@@ -59,7 +57,7 @@ const Home: () => JSX.Element = () => {
             height={['100%', null, '100%', null]}
             borderRadius={6}
           >
-            <AqiGraphScatter sensorDocId={currentSensorDocId} />
+            <AqiGraph sensorDocId={currentSensorDocId} />
           </Box>
         </Flex>
       </Flex>
